+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 
 ## Unreleased
+
+### Added
+- Show hover on reference to default message via number: signature and summary taken from table.
+- Show hover on reference to additional message via number: signature and, if present, a comment on previous line as a summary.
+
 ### Improved
 - Detection of language symbols (prefix: `%`) and quest symbols with prefix `=` or `==`.